Converts correctly Path to FileInfo

+++ files/src/main/java/hr/fer/ilj/workshop/files/FileInfo.java
@@ -2,5 +2,5 @@ package hr.fer.ilj.workshop.files;
 
 import java.nio.file.Path;
 
-public record FileInfo(String name, Path path, int size, FileType type) {
+public record FileInfo(String name, Path path, long size, FileType type) {
 }
